requirements.txt
Django
pytest-django
mock
requests
djangorestframework
pytest